+++ crates/jarl/tests/integration/main.rs
@@ -16,6 +16,7 @@ mod exclude;
 mod help;
 mod helpers;
 mod incompatible_args;
+mod interactive;
 mod jarl;
 mod min_r_version;
 mod no_default_exclude;